+++ packages/pangea-sdk/tests/unit/__init__.py
@@ -1,2 +1,3 @@
 from .test_enum import TestEnums
+from .test_models import TestModels
 from .test_signer import TestSigner